[uploadgui] Move form components
Form component are moved to a separate file to reduce the size of the
form.scala file.
